Fixing delete avatar; adding CSRF protection to delete avatar
Former-commit-id: 512bb4613f9266cc2371dff585d7bd3daac8070f
Former-commit-id: a315dfcdbfeb3076876abcf544495fd6c9f198ef

+++ web/concrete/controllers/single_page/account/avatar.php
@@ -45,12 +45,17 @@ class Avatar extends AccountProfileEditPageController
     public function delete()
     {
         $this->view();
-        $profile = $this->get('profile');
-        $av = $this->get('av');
+        if (!$this->token->validate('delete_avatar')) {
+            $this->error->add($this->token->getErrorMessage());
+        }
+        if (!$this->error->has()) {
+            $profile = $this->get('profile');
+            $av = $this->get('av');
 
-        $service = Core::make('user.avatar');
-        $service->removeAvatar($profile);
-        $this->redirect('/account/avatar', 'deleted');
+            $service = \Core::make('user.avatar');
+            $service->removeAvatar($profile);
+            $this->redirect('/account/avatar', 'deleted');
+        }
     }
 
 }
